% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_advs_params.R
\name{derive_param_map}
\alias{derive_param_map}
\title{Adds a Parameter for Mean Arterial Pressure}
\usage{
derive_param_map()
}
\value{
42
}
\description{
Adds a Parameter for Mean Arterial Pressure
}
\concept{der_prm_bds_findings}
\keyword{der_prm_bds_findings}
